+++ src/applications/kmc-upload-app/create-live/create-live.component.ts
@@ -46,6 +46,7 @@ export class CreateLiveComponent implements OnInit, OnDestroy, AfterViewInit {
   public manualLiveData: ManualLive = {
     name: '',
     description: '',
+    flashHDSURL: '',
     hlsStreamUrl: '',
     useAkamaiHdProtocol: false
   };