+++ src/applications/kmc-upload-app/create-live/manual-live/manual-live.component.ts
@@ -36,6 +36,7 @@ export class ManualLiveComponent implements OnInit, OnDestroy {
       description: this.data.description,
       useAkamaiHdProtocol: this.data.useAkamaiHdProtocol,
       streamUrl: {
+        flashHDSURL: this.data.flashHDSURL,
         hlsStreamUrl: this.data.hlsStreamUrl
       }
     });
@@ -60,6 +61,7 @@ export class ManualLiveComponent implements OnInit, OnDestroy {
   // Create empty structured form on loading
   private _createForm(): void {
     this._streamUrlForm = this._fb.group({
+        flashHDSURL: ['', KalturaValidators.url],
         hlsStreamUrl: ['', KalturaValidators.url]
       },
       {
@@ -83,13 +85,14 @@ export class ManualLiveComponent implements OnInit, OnDestroy {
             name: data.name,
             description: data.description,
             useAkamaiHdProtocol: data.useAkamaiHdProtocol,
+            flashHDSURL: data.streamUrl.flashHDSURL,
             hlsStreamUrl: data.streamUrl.hlsStreamUrl
           });
       });
   }
 
   private _atLeastOneUrlValidator(formgroup: FormGroup) {
-    if (!formgroup.controls['hlsStreamUrl'].value) {
+    if (!formgroup.controls['flashHDSURL'].value && !formgroup.controls['hlsStreamUrl'].value) {
       return {atLeastOneUrl: true};
     } else {
       return null;