+++ src/applications/kmc-upload-app/create-live/manual-live/manual-live.interface.ts
@@ -1,6 +1,7 @@
 export interface ManualLive {
   name: string
   description: string,
+  flashHDSURL: string,
   hlsStreamUrl: string,
   useAkamaiHdProtocol: boolean
 }